Extracting functions into separate files
