+++ src/utils/constants.ts
@@ -1,0 +1,2 @@
+export const ONE_YEAR_IN_DAYS = 365
+export const MINIMUM_AGUINALDO_DAYS_BY_MEXICAN_LAW = 15
